requirements.txt
absl-py==0.7.0
asn1crypto==0.24.0
astor==0.7.1
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.6.1
cryptography==2.5
cycler==0.10.0
Cython==0.29.2
dask==1.0.0
decorator==4.3.0
dlib==19.16.0
docutils==0.14
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
imutils==0.5.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Kivy==1.10.1
Kivy-Garden==0.1.4
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
networkx==2.2
numpy==1.15.0
object-detection==0.1
opencv-python==4.0.0.21
pandas==0.23.4
paramiko==2.4.2
Pillow==5.4.1
protobuf==3.6.1
pyasn1==0.4.5
pycparser==2.19
Pygments==2.3.1
PyNaCl==1.3.0
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
slim==0.1
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
toolz==0.9.0
urllib3==1.24.1
Werkzeug==0.14.1
yattag==1.10.1
